Refactor and updating documentation on AddIdentityProvider
